Changed attribute-equals selector to has-attribute.
The existence of the attribute needs to be
verified, not its contents. Using the has-
attribute selector will allow all of the following
uses to work, where formerly only the first one
would:

	<input autofocus>
	<input autofocus="autofocus" />
	<input autofocus="true">

This obviously does not change any functionality
in the plugin, but points the user in a better
direction and may avoid confusion if a user
happens to employ one of the latter two methods.